backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
pandas==2.1.4
openpyxl==3.1.2
openpyxl-image-loader==1.0.5
pyxlsb==1.0.10
python-multipart==0.0.9
pdfplumber==0.10.3
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
pynput==1.7.6
pyautogui==0.9.54
websockets>=13.0
python-dotenv==1.0.0
